Refactor : 여정 조회 api 함수 분리 및 리팩토링
