+++ client/project-there/src/pages/LayoutPage.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Outlet } from 'react-router-dom';
 import styled from 'styled-components';
 import Header from '../components/common/header/Header';
-import JourneyListProvider from '../contexts/journeyList';
+import JourneyListProvider from '../contexts/JourneyListContext';
 
 const Wrapper = styled.div`
   width: 100vw;
