+++ client/project-there/src/pages/MainPage.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react';
 import JourneyList from '../components/main/JourneyList';
 import Main from '../components/main/Main';
-import JourneyListProvider from '../contexts/journeyList';
+import JourneyListProvider from '../contexts/JourneyListContext';
 import { getAccessToken } from '../lib/utills/controlAccessToken';
 import { useAuthActions } from '../contexts/auth';
 import { baseAxios, authAxios } from '../lib/utills/customAxios';
